- Installazione pacchetto npm "sortablejs"; - Installazione pacchetto composer "bigenergy/laravel-blade-sortable"; - Modifica colonne da float a double dove necessario (per le quantità, principalmente); - Aggiunta colonna "position" alla tabella "item_product"; - Aggiunta la possibilità di posizionare gli elementi che compongono un Articolo tramite drag&drop; - Aggiunta "position" in ItemSeeder.php; - Adattata modale items.show alle nuove modifiche;

+++ database/migrations/2023_04_22_120615_create_location_product_table.php
@@ -12,7 +12,7 @@
 			Schema::create('location_product', function (Blueprint $table) {
 				$table->foreignIdFor(\App\Models\Location::class, 'location_id');
 				$table->foreignIdFor(\App\Models\Product::class, 'product_id');
-				$table->float('quantity');
+				$table->double('quantity');
 			});
 		}
